API rewrite: Updating project structure, bumping version to 0.1.0
