+++ lib/index.js
@@ -2,10 +2,10 @@
 
 module.exports = kmeans;
 
-var initializeKmpp = require('./src/initialize-kmpp');
-var initializeNaive = require('./src/initialize-naive');
-var iterate = require('./src/iterate');
-var distanceMetric = require('./src/l-distance');
+var initializeKmpp = require('./initialize-kmpp');
+var initializeNaive = require('./initialize-naive');
+var iterate = require('./iterate');
+var distanceMetric = require('./distance');
 
 function kmeans (points, opts) {
   var i, k, n, dim, iter, c, initializer, initialize;
