+++ lib/initialize-kmpp.js
