+++ lib/initialize-naive.js
