+++ lib/iterate.js
